third_party/blink/web_tests/wpt_internal/webmcp/select-multiple-events.html

<!DOCTYPE html>
<title>WebMCP: executeTool on &lt;select multiple&gt; dispatches events</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<!-- We'll submit the form to this frame. -->
<iframe name=hidden_iframe style="display:none"></iframe>

<form
  toolname=select_fruits
  tooldescription="Select fruits you like"
  toolautosubmit
  target=hidden_iframe
  action="about:blank">
  <select id=fruits name=fruits multiple>
    <option value=apple>Apple</option>
    <option value=orange>Orange</option>
    <option value=banana>Banana</option>
  </select>
  <input type=submit>
</form>

<script>
promise_test(async t => {
  assert_true(!!navigator.modelContextTesting);

  let fruits = document.getElementById('fruits');
  let inputEventCount = 0;
  let changeEventCount = 0;

  fruits.addEventListener('input', () => {
    inputEventCount++;
  });
  fruits.addEventListener('change', () => {
    changeEventCount++;
  });

  let inputArgs = JSON.stringify({ fruits: ['apple', 'banana'] });

  // This should dispatch events.
  await navigator.modelContextTesting.executeTool('select_fruits', inputArgs);

  assert_true(fruits.options[0].selected);
  assert_false(fruits.options[1].selected);
  assert_true(fruits.options[2].selected);

  assert_equals(inputEventCount, 1);
  assert_equals(changeEventCount, 1);

  // Executing with the same arguments again does not trigger events.
  await navigator.modelContextTesting.executeTool('select_fruits', inputArgs);

  assert_equals(inputEventCount, 1);
  assert_equals(changeEventCount, 1);
}, 'executeTool() on a <select multiple> dispatches input and change events');

</script>
